+++ system/src/documents/FalloutActor.mjs
@@ -1,10 +1,11 @@
 export default class FalloutActor extends Actor {
 
+	isSleeping = false;
+
 	constructor(object, options={}) {
 		super(object, options);
-
-		this.isSleeping = false;
 	}
+
 
 	/**
 	 * Update any settlement sheets that may be linked to the deleted Actor
@@ -50,6 +51,11 @@ export default class FalloutActor extends Actor {
 		totals.diff = totals.max - totals.total;
 
 		return totals;
+	}
+
+
+	get isAlcoholic() {
+		return this.system.conditions.alcoholic;
 	}
 
 
@@ -171,7 +177,6 @@ export default class FalloutActor extends Actor {
 		// things organized.
 
 		if (this.type === "character" || this.type === "robot") {
-			this._calculateAttributes();
 			this._calculateDefense();
 			this._calculateInitiative();
 			this._calculateMaxHp();
@@ -366,13 +371,6 @@ export default class FalloutActor extends Actor {
 		}
 		// ADD OUTFITED LIST FOR DISPLAY
 		this.system.outfittedLocations = outfittedLocations;
-	}
-
-	_calculateAttributes() {
-		for (const key in this.system.attributes) {
-			const attribute = this.system.attributes[key];
-			attribute.current = attribute.value + attribute.bonus;
-		}
 	}
 
 	_calculateDefense() {
@@ -1380,13 +1378,19 @@ export default class FalloutActor extends Actor {
 			}
 
 			if (consumableType === "chem" && item.system.addictive) {
-				const alreadyAddicted = await this.isAddictedToChem(item);
+				const addictionName = item.system.consumableGroup !== ""
+					? item.system.consumableGroup
+					: item.name;
+
+				const alreadyAddicted = await this.isAddictedToChem(addictionName);
 
 				const chemId = item.name.slugify();
 				const dosageKey = `system.chemDoses.${chemId}`;
 
+				let scenes = item.system.duration === "lasting" ? 2 : 1;
+
 				let	newDosage = this.system.chemDoses[chemId]?.doses ?? 0;
-				newDosage++;
+				newDosage = alreadyAddicted ? 1 : newDosage + 1;
 
 				const addictionNumberExceeded =
 					newDosage >= item.system.addiction;
@@ -1400,16 +1404,30 @@ export default class FalloutActor extends Actor {
 					fallout.Roller2D20.showDiceSoNice(addictedRoll);
 
 					if (parseInt(roll.result) >= item.system.addiction) {
+						newDosage = 1;
 
-						const addictionName = item.system.consumableGroup !== ""
-							? item.system.consumableGroup
-							: item.name;
+						// Automatically add Addiction item of the correct type
+						// to the character
+						//
+						const addiction = (
+							await fallout.compendiums.addictions()
+						).find(
+							a => a.name === addictionName
+						);
+
+						if (addiction) {
+							this.createEmbeddedDocuments("Item", [addiction]);
+						}
+						else {
+							fallout.logger.warn(`Unable to fund addiction with the name ${addictionName}`);
+						}
 
 						fallout.chat.renderGeneralMessage(
 							this,
 							{
 								title: game.i18n.localize("FALLOUT.CHAT_MESSAGE.addiction.title"),
-								body: game.i18n.format("FALLOUT.CHAT_MESSAGE.addiction.body",
+								body: game.i18n.format(
+									"FALLOUT.CHAT_MESSAGE.addiction.body",
 									{
 										actorName: this.name,
 										itemName: addictionName,
@@ -1423,9 +1441,11 @@ export default class FalloutActor extends Actor {
 
 				actorUpdateData[dosageKey] = {
 					addiction: item.system.addiction,
+					addictionName,
 					doses: newDosage,
 					id: chemId,
 					name: item.name,
+					scenes,
 				};
 
 			}
@@ -1586,15 +1606,11 @@ export default class FalloutActor extends Actor {
 		);
 	}
 
-	async isAddictedToChem(item) {
-		const chemName = item.system.consumableGroup !== ""
-			? item.system.consumableGroup
-			: item.name;
-
+	async isAddictedToChem(addictionName) {
 		const addiction = this.items.filter(
 			i => i.type === "addiction"
 		).find(
-			i => i.name === chemName
+			i => i.name === addictionName
 		);
 
 		return addiction ? true : false;
@@ -1648,6 +1664,39 @@ export default class FalloutActor extends Actor {
 				"system.quantity": newQuantity,
 			}]);
 		}
+	}
+
+	async updateAddictions() {
+		const updateData = {};
+
+		if (this.isAlcoholic) {
+			updateData["system.conditions.intoxication"] = 0;
+		}
+
+		for (const doseKey in this.system.chemDoses) {
+			const dose = this.system.chemDoses[doseKey];
+
+			const isAddicted = await this.isAddictedToChem(
+				dose.addictionName
+			);
+
+			const scenesRemaining = dose.scenes - 1;
+
+			if (isAddicted) {
+				if (scenesRemaining <= 0) {
+					updateData[`system.chemDoses.-=${doseKey}`] = null;
+				}
+				else {
+					dose.scenes = scenesRemaining;
+					updateData[`system.chemDoses.${doseKey}`] = dose;
+				}
+			}
+			else {
+				updateData[`system.chemDoses.${doseKey}`] = dose;
+			}
+		}
+
+		this.update(updateData);
 	}
 
 	async resetChemDoses() {
@@ -1715,6 +1764,8 @@ export default class FalloutActor extends Actor {
 			"system.conditions.sleep": newSleepStatus,
 			"system.conditions.wellRested": newWellRested,
 		};
+
+		await this.updateAddictions();
 
 		if (newWellRested && this.isNotWellRested) {
 			updateData["system.health.value"] = this.system.health.value + 2;